+++ lib/asset_sync.rb
@@ -1,6 +1,3 @@
-require 'fog/aws'
-require 'active_model'
-require 'erb'
 require "asset_sync/asset_sync"
 require 'asset_sync/config'
 require 'asset_sync/storage'