+++ lib/asset_sync/config.rb
@@ -1,3 +1,9 @@
+# frozen_string_literal: true
+
+require "active_model"
+require "erb"
+require "yaml"
+
 module AssetSync
   class Config
     include ActiveModel::Validations
@@ -13,26 +19,39 @@ module AssetSync
     attr_accessor :always_upload
     attr_accessor :ignored_files
     attr_accessor :prefix
-    attr_accessor :public_path
     attr_accessor :enabled
     attr_accessor :custom_headers
     attr_accessor :run_on_precompile
     attr_accessor :invalidate
     attr_accessor :cdn_distribution_id
+    attr_accessor :cache_asset_regexps
+    attr_accessor :include_manifest
+    attr_writer :public_path
 
     # FOG configuration
     attr_accessor :fog_provider          # Currently Supported ['AWS', 'Rackspace']
     attr_accessor :fog_directory         # e.g. 'the-bucket-name'
     attr_accessor :fog_region            # e.g. 'eu-west-1'
+    attr_reader   :fog_public            # e.g. true, false, "default"
 
     # Amazon AWS
-    attr_accessor :aws_access_key_id, :aws_secret_access_key, :aws_reduced_redundancy, :aws_iam_roles
+    attr_accessor :aws_access_key_id, :aws_secret_access_key, :aws_reduced_redundancy, :aws_iam_roles, :aws_signature_version
+    attr_accessor :fog_host              # e.g. 's3.amazonaws.com'
+    attr_accessor :fog_port              # e.g. '9000'
+    attr_accessor :fog_path_style        # e.g. true
+    attr_accessor :fog_scheme            # e.g. 'http'
 
     # Rackspace
     attr_accessor :rackspace_username, :rackspace_api_key, :rackspace_auth_url
 
     # Google Storage
-    attr_accessor :google_storage_secret_access_key, :google_storage_access_key_id
+    attr_accessor :google_storage_secret_access_key, :google_storage_access_key_id  # when using S3 interop
+    attr_accessor :google_json_key_location # when using service accounts
+    attr_accessor :google_project # when using service accounts
+
+    # Azure Blob with Fog::AzureRM
+    attr_accessor :azure_storage_account_name
+    attr_accessor :azure_storage_access_key
 
     validates :existing_remote_files, :inclusion => { :in => %w(keep delete ignore) }
 
@@ -43,11 +62,14 @@ module AssetSync
     validates :aws_secret_access_key, :presence => true, :if => proc {aws? && !aws_iam?}
     validates :rackspace_username,    :presence => true, :if => :rackspace?
     validates :rackspace_api_key,     :presence => true, :if => :rackspace?
-    validates :google_storage_secret_access_key,  :presence => true, :if => :google?
-    validates :google_storage_access_key_id,      :presence => true, :if => :google?
+    validates :google_storage_secret_access_key,  :presence => true, :if => :google_interop?
+    validates :google_storage_access_key_id,      :presence => true, :if => :google_interop?
+    validates :google_json_key_location,          :presence => true, :if => :google_service_account?
+    validates :google_project,                    :presence => true, :if => :google_service_account?
 
     def initialize
       self.fog_region = nil
+      self.fog_public = true
       self.existing_remote_files = 'keep'
       self.gzip_compression = false
       self.manifest = false
@@ -60,12 +82,16 @@ module AssetSync
       self.run_on_precompile = true
       self.cdn_distribution_id = nil
       self.invalidate = []
-      load_yml! if defined?(Rails) && yml_exists?
+      self.cache_asset_regexps = []
+      self.include_manifest = false
+      @additional_local_file_paths_procs = []
+
+      load_yml! if defined?(::Rails) && yml_exists?
     end
 
     def manifest_path
       directory =
-        Rails.application.config.assets.manifest || default_manifest_directory
+        ::Rails.application.config.assets.manifest || default_manifest_directory
       File.join(directory, "manifest.yml")
     end
 
@@ -78,7 +104,7 @@ module AssetSync
     end
 
     def aws?
-      fog_provider == 'AWS'
+      fog_provider =~ /aws/i
     end
 
     def aws_rrs?
@@ -94,7 +120,7 @@ module AssetSync
     end
 
     def log_silently?
-      ENV['RAILS_GROUPS'] == 'assets' || self.log_silently == false
+      !!self.log_silently
     end
 
     def enabled?
@@ -102,50 +128,72 @@ module AssetSync
     end
 
     def rackspace?
-      fog_provider == 'Rackspace'
+      fog_provider =~ /rackspace/i
     end
 
     def google?
-      fog_provider == 'Google'
+      fog_provider =~ /google/i
+    end
+
+    def google_interop?
+      google? && google_json_key_location.nil?
+    end
+
+    def google_service_account?
+      google? && google_json_key_location
+    end
+
+    def azure_rm?
+      fog_provider =~ /azurerm/i
+    end
+
+    def cache_asset_regexp=(cache_asset_regexp)
+      self.cache_asset_regexps = [cache_asset_regexp]
     end
 
     def yml_exists?
-      defined?(Rails.root) ? File.exist?(self.yml_path) : false
+      defined?(::Rails.root) ? File.exist?(self.yml_path) : false
     end
 
     def yml
-      @yml ||= YAML.load(ERB.new(IO.read(yml_path)).result)[Rails.env] || {}
-    rescue Psych::SyntaxError
-      @yml = {}
+      @yml ||= ::YAML.load(::ERB.new(IO.read(yml_path)).result)[::Rails.env] || {}
     end
 
     def yml_path
-      Rails.root.join("config", "asset_sync.yml").to_s
+      ::Rails.root.join("config", "asset_sync.yml").to_s
     end
 
     def assets_prefix
       # Fix for Issue #38 when Rails.config.assets.prefix starts with a slash
-      self.prefix || Rails.application.config.assets.prefix.sub(/^\//, '')
+      self.prefix || ::Rails.application.config.assets.prefix.sub(/^\//, '')
     end
 
     def public_path
-      @public_path || Rails.public_path
+      @public_path || ::Rails.public_path
     end
 
     def load_yml!
       self.enabled                = yml["enabled"] if yml.has_key?('enabled')
       self.fog_provider           = yml["fog_provider"]
+      self.fog_host               = yml["fog_host"]
+      self.fog_port               = yml["fog_port"]
       self.fog_directory          = yml["fog_directory"]
       self.fog_region             = yml["fog_region"]
+      self.fog_public             = yml["fog_public"] if yml.has_key?("fog_public")
+      self.fog_path_style         = yml["fog_path_style"]
+      self.fog_scheme             = yml["fog_scheme"]
       self.aws_access_key_id      = yml["aws_access_key_id"]
       self.aws_secret_access_key  = yml["aws_secret_access_key"]
       self.aws_reduced_redundancy = yml["aws_reduced_redundancy"]
       self.aws_iam_roles          = yml["aws_iam_roles"]
+      self.aws_signature_version  = yml["aws_signature_version"]
       self.rackspace_username     = yml["rackspace_username"]
       self.rackspace_auth_url     = yml["rackspace_auth_url"] if yml.has_key?("rackspace_auth_url")
       self.rackspace_api_key      = yml["rackspace_api_key"]
-      self.google_storage_secret_access_key = yml["google_storage_secret_access_key"]
-      self.google_storage_access_key_id     = yml["google_storage_access_key_id"]
+      self.google_json_key_location = yml["google_json_key_location"] if yml.has_key?("google_json_key_location")
+      self.google_project = yml["google_project"] if yml.has_key?("google_project")
+      self.google_storage_secret_access_key = yml["google_storage_secret_access_key"] if yml.has_key?("google_storage_secret_access_key")
+      self.google_storage_access_key_id     = yml["google_storage_access_key_id"] if yml.has_key?("google_storage_access_key_id")
       self.existing_remote_files  = yml["existing_remote_files"] if yml.has_key?("existing_remote_files")
       self.gzip_compression       = yml["gzip_compression"] if yml.has_key?("gzip_compression")
       self.manifest               = yml["manifest"] if yml.has_key?("manifest")
@@ -156,6 +204,11 @@ module AssetSync
       self.run_on_precompile      = yml["run_on_precompile"] if yml.has_key?("run_on_precompile")
       self.invalidate             = yml["invalidate"] if yml.has_key?("invalidate")
       self.cdn_distribution_id    = yml['cdn_distribution_id'] if yml.has_key?("cdn_distribution_id")
+      self.cache_asset_regexps    = yml['cache_asset_regexps'] if yml.has_key?("cache_asset_regexps")
+      self.include_manifest       = yml['include_manifest'] if yml.has_key?("include_manifest")
+
+      self.azure_storage_account_name = yml['azure_storage_account_name'] if yml.has_key?("azure_storage_account_name")
+      self.azure_storage_access_key   = yml['azure_storage_access_key'] if yml.has_key?("azure_storage_access_key")
 
       # TODO deprecate the other old style config settings. FML.
       self.aws_access_key_id      = yml["aws_access_key"] if yml.has_key?("aws_access_key")
@@ -186,32 +239,130 @@ module AssetSync
             :aws_secret_access_key => aws_secret_access_key
           })
         end
+        options.merge!({:host => fog_host}) if fog_host
+        options.merge!({:port => fog_port}) if fog_port
+        options.merge!({:scheme => fog_scheme}) if fog_scheme
+        options.merge!({:aws_signature_version => aws_signature_version}) if aws_signature_version
+        options.merge!({:path_style => fog_path_style}) if fog_path_style
+        options.merge!({:region => fog_region}) if fog_region
       elsif rackspace?
         options.merge!({
           :rackspace_username => rackspace_username,
           :rackspace_api_key => rackspace_api_key
         })
-        options.merge!({
-          :rackspace_region => fog_region
-        }) if fog_region
+        options.merge!({ :rackspace_region => fog_region }) if fog_region
         options.merge!({ :rackspace_auth_url => rackspace_auth_url }) if rackspace_auth_url
       elsif google?
+        if google_json_key_location
+          options.merge!({:google_json_key_location => google_json_key_location, :google_project => google_project})
+        else
+          options.merge!({
+            :google_storage_secret_access_key => google_storage_secret_access_key,
+            :google_storage_access_key_id => google_storage_access_key_id
+          })
+        end
+        options.merge!({:region => fog_region}) if fog_region
+      elsif azure_rm?
+        require 'fog/azurerm'
         options.merge!({
-          :google_storage_secret_access_key => google_storage_secret_access_key,
-          :google_storage_access_key_id => google_storage_access_key_id
+          :azure_storage_account_name => azure_storage_account_name,
+          :azure_storage_access_key   => azure_storage_access_key,
         })
+        options.merge!({:environment => fog_region}) if fog_region
       else
-        raise ArgumentError, "AssetSync Unknown provider: #{fog_provider} only AWS and Rackspace are supported currently."
+        raise ArgumentError, "AssetSync Unknown provider: #{fog_provider} only AWS, Rackspace and Google are supported currently."
       end
 
-      options.merge!({:region => fog_region}) if fog_region && !rackspace?
-      return options
+      options
+    end
+
+    # @api
+    def add_local_file_paths(&block)
+      @additional_local_file_paths_procs =
+        additional_local_file_paths_procs + [block]
+    end
+
+    # @api private
+    #   This is to be called in Storage
+    #   Not to be called by user
+    def additional_local_file_paths
+      return [] if additional_local_file_paths_procs.empty?
+
+      # Using `Array()` to ensure it works when single value is returned
+      additional_local_file_paths_procs.each_with_object([]) do |proc, paths|
+        paths.concat(Array(proc.call))
+      end
+    end
+
+    #@api
+    def file_ext_to_mime_type_overrides
+      @file_ext_to_mime_type_overrides ||= FileExtToMimeTypeOverrides.new
+    end
+
+    def fog_public=(new_val)
+      @fog_public = FogPublicValue.new(new_val)
     end
 
   private
 
+    # This is a proc to get additional local files paths
+    # Since this is a proc it won't be able to be configured by a YAML file
+    attr_reader :additional_local_file_paths_procs
+
     def default_manifest_directory
-      File.join(Rails.public_path, assets_prefix)
+      File.join(::Rails.public_path, assets_prefix)
+    end
+
+
+    # @api private
+    class FileExtToMimeTypeOverrides
+      def initialize
+        # The default is to prevent new mime type `application/ecmascript` to be returned
+        # which disables compression on some CDNs
+        @overrides = {
+          "js" => "application/javascript",
+        }
+      end
+
+      # @api
+      def add(ext, mime_type)
+        # Symbol / Mime type object might be passed in
+        # But we want strings only
+        @overrides.store(
+          ext.to_s, mime_type.to_s,
+        )
+      end
+
+      # @api
+      def clear
+        @overrides = {}
+      end
+
+
+      # @api private
+      def key?(key)
+        @overrides.key?(key)
+      end
+
+      # @api private
+      def fetch(key)
+        @overrides.fetch(key)
+      end
+    end
+
+    # @api private
+    class FogPublicValue
+      def initialize(val)
+        @value = val
+      end
+
+      def use_explicit_value?
+        @value.to_s != "default"
+      end
+
+      def to_bool
+        !!@value
+      end
     end
   end
 end