+++ lib/asset_sync/multi_mime.rb
@@ -4,14 +4,21 @@ module AssetSync
   class MultiMime
 
     def self.lookup(ext)
+      overrides =
+        ::AssetSync.config.file_ext_to_mime_type_overrides
+      if overrides.key?(ext)
+        return overrides.fetch(ext)
+      end
 
-      if defined?(Mime::Type)
-        Mime::Type.lookup_by_extension(ext)
-      elsif defined?(Rack::Mime)
+      if defined?(::MIME::Types)
+        ::MIME::Types.type_for(ext).first.to_s
+      elsif defined?(::Mime::Type)
+        ::Mime::Type.lookup_by_extension(ext).to_s
+      elsif defined?(::Rack::Mime)
         ext_with_dot = ".#{ext}"
-        Rack::Mime.mime_type(ext_with_dot)
+        ::Rack::Mime.mime_type(ext_with_dot)
       else
-        ::MIME::Types.type_for(ext).first
+        raise "No library found for mime type lookup"
       end
 
     end