+++ spec/integration/azure_rm_integration_spec.rb
@@ -1,12 +1,13 @@
 require File.dirname(__FILE__) + '/../spec_helper'
-require "fog/aws"
+require "fog/azurerm"
 
 def bucket(name)
   options = {
-    :provider => 'AWS',
-    :aws_access_key_id => ENV['AWS_ACCESS_KEY_ID'],
-    :aws_secret_access_key => ENV['AWS_SECRET_ACCESS_KEY']
+    :provider => 'AzureRM',
+    :azure_storage_account_name => ENV['AZURE_STORAGE_ACCOUNT_NAME'],
+    :azure_storage_access_key => ENV['AZURE_STORAGE_ACCESS_KEY']
   }
+  options.merge!({ :environment => ENV['FOG_REGION'] }) if ENV.has_key?('FOG_REGION')
 
   connection = Fog::Storage.new(options)
   connection.directories.get(ENV['FOG_DIRECTORY'], :prefix => name)
@@ -47,14 +48,11 @@ describe "AssetSync" do
 
     files = bucket(@prefix).files
 
-    app_js_path = files.select{ |f| f.key =~ app_js_regex }.first
-    app_js_gz_path = files.select{ |f| f.key =~ app_js_gz_regex }.first
+    app_js = files.select{ |f| f.key =~ app_js_regex }.first
+    expect(app_js.content_type).to eq("application/javascript")
 
-    app_js = files.get( app_js_path.key )
-    expect(app_js.content_type).to eq("text/javascript")
-
-    app_js_gz = files.get( app_js_gz_path.key )
-    expect(app_js_gz.content_type).to eq("text/javascript")
+    app_js_gz = files.select{ |f| f.key =~ app_js_gz_regex }.first
+    expect(app_js_gz.content_type).to eq("application/javascript")
     expect(app_js_gz.content_encoding).to eq("gzip")
   end
 
@@ -69,7 +67,7 @@ describe "AssetSync" do
 
     app_js_path = files.select{ |f| f.key =~ app_js_regex }.first
     app_js = files.get( app_js_path.key )
-    expect(app_js.content_type).to eq("text/javascript")
+    expect(app_js.content_type).to eq("application/javascript")
   end
 
 end